Add support for file-specified delimiters, and ODS files.
File-specified delimiters:
In LibTehthu's native plaintext file format, you can specify the delimiter
to be used in the parsing as one of the first two lines. Just give the delim-
iter as a single-character line which is either the first or second line.

ODS files:
Support OASIS OpenDocument standard Spreadsheet files as produced by
Microsoft Excel 2007 / 2010 or OpenOffice.org 2.0 or later.

See README.txt for details on ODS support.